requirements.txt
flask==3.0.0
torch==2.1.0
transformers==4.35.0
pillow==10.0.0
google-generativeai==0.3.0
flask-cors==4.0.0